move EventCard to local for QTUMPrediction
