+++ src/scenes/QtumPrediction/components/EventCard/styles.js
