+++ src/scenes/QtumPrediction/components/TopActions/index.js
