+++ src/scenes/QtumPrediction/components/TopActions/styles.js
